# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Code4Me V2 API",
    "description": "The complete API for Code4Me V2",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/completion/request": {
      "post": {
        "tags": [
          "Completion"
        ],
        "summary": "Request Completion",
        "description": "Handle a code completion request.\n\nSteps:\n- Authenticate session, user, and project tokens via Redis.\n- Redact any secrets from the context before processing.\n- Prepare optional Celery tasks for storing context and telemetry.\n- Aggregate multi-file context if available.\n- Run completion models concurrently using a thread pool.\n- Queue database update tasks asynchronously using Celery.\n- Return completion results or appropriate error responses.",
        "operationId": "request_completion_api_completion_request_post",
        "parameters": [
          {
            "name": "session_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Session Token"
            }
          },
          {
            "name": "project_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Project Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCompletion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionPostResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/InvalidOrExpiredSessionToken"
                    },
                    {
                      "$ref": "#/components/schemas/InvalidOrExpiredProjectToken"
                    }
                  ],
                  "title": "Response 401 Request Completion Api Completion Request Post"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateCompletionsError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/completion/{query_id}": {
      "get": {
        "tags": [
          "Completion"
        ],
        "summary": "Get Completions By Query",
        "description": "Retrieve code completions associated with a specific query ID.\n\nThis endpoint validates the user's session token and authorization,\nensures the requested query exists and belongs to the requesting user,\nand then fetches all the associated completion generations from the database.\n\nParameters:\n- query_id (UUID): The unique identifier for the meta query whose completions are requested.\n- app (App, dependency): The application instance providing database and Redis access.\n- session_token (str, cookie): The session token cookie for user authentication.\n\nReturns:\n- JsonResponseWithStatus: JSON response with HTTP status and either the completions data\n  or an error response detailing the failure reason.\n\nPossible responses:\n- 200: Successfully retrieved completions for the given query ID.\n- 401: Session token is invalid, expired, or missing.\n- 403: User does not have access rights to the requested query.\n- 404: The requested query ID does not exist.\n- 422, 429: Various client errors (validation or rate limiting).\n- 500: Internal server error while retrieving completions.",
        "operationId": "get_completions_by_query_api_completion__query_id__get",
        "parameters": [
          {
            "name": "query_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Query Id"
            }
          },
          {
            "name": "session_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionPostResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidOrExpiredSessionToken"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoAccessToGetQueryError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryNotFoundError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrieveCompletionsError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/chat/request": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Request Chat Completion",
        "description": "Request chat completions based on provided messages.\n\nThe contract here is that the request always contains all the history of the chat as well\n\nWe do this because it could be that the user modifies the chat history in the frontend\nand we want to ensure that the chat completions are based on the latest state of the chat.\n\nTake the case where the user edits a message in the chat history midway through a chat.\nWe don't want to be generating a completion based on the old state of the chat.",
        "operationId": "request_chat_completion_api_chat_request_post",
        "parameters": [
          {
            "name": "session_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Session Token"
            }
          },
          {
            "name": "project_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Project Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestChatCompletion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatHistoryResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/InvalidOrExpiredSessionToken"
                    },
                    {
                      "$ref": "#/components/schemas/InvalidOrExpiredProjectToken"
                    }
                  ],
                  "title": "Response 401 Request Chat Completion Api Chat Request Post"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateChatCompletionsError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatHistoryResponse": {
        "properties": {
          "chat_id": {
            "type": "string",
            "format": "uuid",
            "title": "Chat Id",
            "description": "Unique chat session ID"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Title of the chat"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ChatHistoryItem"
            },
            "type": "array",
            "title": "History",
            "description": "List of message exchanges"
          }
        },
        "type": "object",
        "required": [
          "chat_id",
          "title",
          "history"
        ],
        "title": "ChatHistoryResponse",
        "description": "Complete chat conversation history for a session."
      },
...
      "CompletionPostResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Completions generated successfully. Ready for display in your IDE."
          },
          "data": {
            "$ref": "#/components/schemas/ResponseCompletionResponseData",
            "description": "Generated code completions including query ID and suggestions"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "CompletionPostResponse"
      },
...
      "ErrorResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Error message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ErrorResponse"
      },
...
      "GenerateChatCompletionsError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Server failed to generate chat completions."
          }
        },
        "type": "object",
        "title": "GenerateChatCompletionsError"
      },
      "GenerateCompletionsError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Server failed to generate completions."
          }
        },
        "type": "object",
        "title": "GenerateCompletionsError"
      },
...
      "InvalidOrExpiredProjectToken": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Invalid or expired project token. Acquire a session token first."
          }
        },
        "type": "object",
        "title": "InvalidOrExpiredProjectToken"
      },
      "InvalidOrExpiredSessionToken": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Session not found! You are not authenticated or your session has expired. Login before you can perform this action."
          }
        },
        "type": "object",
        "title": "InvalidOrExpiredSessionToken"
      },
...
      "NoAccessToGetQueryError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "You do not have permission to access this query."
          }
        },
        "type": "object",
        "title": "NoAccessToGetQueryError"
      },
...
      "QueryChatMessageRole": {
        "type": "string",
        "enum": [
          "user",
          "assistant",
          "system"
        ],
        "title": "QueryChatMessageRole"
      },
      "QueryNotFoundError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Query not found"
          }
        },
        "type": "object",
        "title": "QueryNotFoundError"
      },
      "RequestChatCompletion": {
        "properties": {
          "model_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Model Ids",
            "description": "Models to use for chat completion"
          },
          "chat_id": {
            "type": "string",
            "format": "uuid",
            "title": "Chat Id",
            "description": "Chat ID"
          },
          "messages": {
            "items": {
              "prefixItems": [
                {
                  "$ref": "#/components/schemas/QueryChatMessageRole"
                },
                {
                  "type": "string"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Messages",
            "description": "Chat messages as a list of tuples"
          },
          "context": {
            "$ref": "#/components/schemas/ContextData",
            "description": "Context data for completion"
          },
          "store_context": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Store Context",
            "description": "Whether to store context data in database",
            "default": false
          },
          "contextual_telemetry": {
            "$ref": "#/components/schemas/ContextualTelemetryData",
            "description": "Contextual telemetry data"
          },
          "store_contextual_telemetry": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Store Contextual Telemetry",
            "description": "Whether to store contextual telemetry in database",
            "default": true
          },
          "behavioral_telemetry": {
            "$ref": "#/components/schemas/BehavioralTelemetryData",
            "description": "Behavioral telemetry data"
          },
          "store_behavioral_telemetry": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Store Behavioral Telemetry",
            "description": "Whether to store behavioral telemetry in database",
            "default": true
          },
          "web_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Web Enabled",
            "description": "Whether web access is enabled",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "model_ids",
          "chat_id",
          "messages",
          "context",
          "contextual_telemetry",
          "behavioral_telemetry"
        ],
        "title": "RequestChatCompletion"
      },
      "RequestCompletion": {
        "properties": {
          "model_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Model Ids",
            "description": "Models to use for completion"
          },
          "context": {
            "$ref": "#/components/schemas/ContextData",
            "description": "Context data for completion"
          },
          "store_context": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Store Context",
            "description": "Whether to store context data in database",
            "default": false
          },
          "contextual_telemetry": {
            "$ref": "#/components/schemas/ContextualTelemetryData",
            "description": "Contextual telemetry data"
          },
          "store_contextual_telemetry": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Store Contextual Telemetry",
            "description": "Whether to store contextual telemetry in database",
            "default": true
          },
          "behavioral_telemetry": {
            "$ref": "#/components/schemas/BehavioralTelemetryData",
            "description": "Behavioral telemetry data"
          },
          "store_behavioral_telemetry": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Store Behavioral Telemetry",
            "description": "Whether to store behavioral telemetry in database",
            "default": true
          },
          "stop_sequences": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Sequences",
            "description": "List of sequences that signal the end of generation",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "model_ids",
          "context",
          "contextual_telemetry",
          "behavioral_telemetry"
        ],
        "title": "RequestCompletion"
      },
...
      "RetrieveCompletionsError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Server failed to retrieve completions."
          }
        },
        "type": "object",
        "title": "RetrieveCompletionsError"
      },
...
    }
  }
}